aaa
aaa

+++ src/model/Centro.java
@@ -2,23 +2,23 @@ package model;
 
 public class Centro {
 
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	
 	public Centro() {}
 	
-	public Centro(int x, int y) {
+	public Centro(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	public void setY(int y) {
